requirements.txt
# requirements.txt
# Updated with specific package versions to ensure compatibility.
# Pin Flask and Werkzeug to versions compatible with flask-mongoengine
# to solve the JSONEncoder import error.
Flask==2.2.2
Werkzeug==2.2.2
python-dotenv
click
Flask-WTF
flask-mongoengine
PyJWT
waitress